enter claimInfo first, then visit info

+++ migrations/20190919090553-create-billing-info.js
@@ -22,7 +22,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       value: {
-        type: Sequelize.FLOAT
+        type: Sequelize.BIGINT
       },
       createdAt: {
         allowNull: false,
